pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zagpy"
version = "0.1.0"
description = "A simple, yet flexible, RAG framework."
readme = "LICENSE"
requires-python = ">=3.10"
dependencies = [
    # Core dependencies
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=0.19.0",
    
    # Utilities
    "tiktoken>=0.7.0",
    "tenacity>=8.0.0",
    
    # Data processing
    "pandas>=1.3.0",
    
    # Vector storage
    "chromadb>=0.5.0",
    "qdrant-client>=1.16.2",
    "lancedb>=0.15.0",
    "pymilvus>=2.4.0",
    
    # Full-text search
    "meilisearch>=0.31.0",
    
    # LLM providers
    "openai>=1.0.0",
    "chakpy>=0.2.0",
    "ollama>=0.4.0",
    
    # Embedding models
    "sentence-transformers>=2.0.0",
    
    # Document readers
    "markitdown>=0.0.1",
    "docling>=2.0.0",
    "mineru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["zag*"]
